moving the middleware and creating a wrapper class for phalcon request for later module rest route implementation
