+++ src/Clarity/Support/Phalcon/Http/Request.php
@@ -1,0 +1,8 @@
+<?php
+namespace Clarity\Support\Phalcon\Http;
+
+use Phalcon\Http\Request as BaseRequest;
+
+class Request extends BaseRequest
+{
+}
