--- src/Clarity/composer.json
@@ -1,27 +1,0 @@
-{
-    "name": "clarity/kernel",
-    "description": "",
-    "license": "MIT",
-    "homepage": "http://github.com/ps-clarity/kernel",
-    "support": {
-        "issues": "https://github.com/ps-clarity/kernel/issues",
-        "source": "https://github.com/ps-clarity/kernel"
-    },
-    "authors": [
-        {
-            "name": "Daison Carino",
-            "email": "daison12006013@gmail.com"
-        }
-    ],
-    "require": {
-        "clarity/support": "dev-master",
-        "symfony/console": "^2.7",
-        "league/flysystem": "^1.0",
-        "monolog/monolog": "^1.15"
-    },
-    "autoload": {
-        "psr-4": {
-            "Clarity\\": ""
-        }
-    }
-}
